move and rename package containing sslContextProvider
